+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/views/cpd/CPDViewTooltipListener.java
@@ -19,7 +19,8 @@ import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.Tree;
 import org.eclipse.swt.widgets.TreeItem;
 
-import net.sourceforge.pmd.cpd.Match;
+import net.sourceforge.pmd.cpd.Mark;
+import net.sourceforge.pmd.eclipse.runtime.cmd.internal.CpdMatchWithSourceCode;
 
 /**
  * This class is created for showing the source code of a tablerow via a
@@ -89,8 +90,8 @@ public class CPDViewTooltipListener implements Listener {
         case SWT.MouseHover:
             if (item != null) {
                 final TreeNode node = (TreeNode) item.getData();
-                if (node.getValue() instanceof Match) {
-                    final Match match = (Match) node.getValue();
+                if (node.getValue() instanceof CpdMatchWithSourceCode) {
+                    final CpdMatchWithSourceCode match = (CpdMatchWithSourceCode) node.getValue();
                     final String text = createText(match);
                     showTooltip(item, text);
                 }
@@ -102,8 +103,9 @@ public class CPDViewTooltipListener implements Listener {
         }
     }
 
-    private String createText(Match match) {
-        final String text = match.getSourceCodeSlice().replaceAll("\t", "    ");
+    private String createText(CpdMatchWithSourceCode match) {
+        Mark firstMark = match.getMatch().getFirstMark();
+        final String text = match.getSourceCodeSlices().get(firstMark).toString().replaceAll("\t", "    ");
         final StringBuilder outputString = new StringBuilder();
         final StringTokenizer lines = new StringTokenizer(text, "\n");
         for (int i = 0; lines.hasMoreTokens(); i++) {